#Jira BRS-3707 | [Api] Bookkeep using a invoiceReportLayoutId doesn't work - refactor
