+++ src/Xena.Contracts/Domain/Abstract/TransactionalDto.cs
@@ -1,0 +1,8 @@
+﻿namespace Xena.Contracts.Domain
+{
+    public abstract class TransactionalDto : FiscalDto, IHasIdDto
+    {
+        public long? Id { get; set; }
+        public int Version { get; set; }
+    }
+}
