+++ src/Xena.Contracts/Domain/Interfaces/IEntityDto.cs
@@ -1,0 +1,8 @@
+﻿namespace Xena.Contracts.Domain
+{
+    public interface IEntityDto : IHasIdDto
+    {
+        int Version { get; }
+        bool IsDeactivated { get; }
+    }
+}
