+++ src/Xena.Contracts/Domain/Interfaces/IHasIdDto.cs
@@ -1,0 +1,7 @@
+﻿namespace Xena.Contracts.Domain
+{
+    public interface IHasIdDto
+    {
+        long? Id { get; }
+    }
+}
