+++ src/Xena.Contracts/Domain/Interfaces/IHasUserIdDto.cs
@@ -1,0 +1,7 @@
+﻿namespace Xena.Contracts.Domain
+{
+    public interface IHasUserIdDto
+    {
+        long UserId { get; set; }
+    }
+}
