+++ src/Xena.Contracts/Domain/Interfaces/IHasXSLLayoutDto.cs
@@ -1,0 +1,11 @@
+﻿using System.Globalization;
+
+namespace Xena.Contracts.Domain
+{
+    // public interface IHasXSLLayout : IHasIdDto
+    // {
+    //     string Group { get; }
+    //     string XSL { get; }
+    //     CultureInfo Culture { get; }
+    // }
+}
