+++ src/Xena.Contracts/Domain/ReportLayoutRelated/ReportLayoutDto.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.ComponentModel;
-using Xena.Contracts.Domain;
 using Xena.Common.ExtensionMethods;
 
-namespace Xena.Contracts.Helpers
+namespace Xena.Contracts.Domain
 {
     public class ReportLayoutDto : EntityDto
     {
