+++ src/Xena.Contracts/Domain/ReportLayoutRelated/ReportLayoutTextDto.cs
@@ -1,6 +1,4 @@
-﻿using Xena.Contracts.Domain;
-
-namespace Xena.Contracts.Helpers
+﻿namespace Xena.Contracts.Domain
 {
     public class ReportLayoutTextDto : EntityDto
     {
